refactor piewidget and extract some logic to a hook
