+++ packages/react-ui/src/widgets/PieWidgetUI/PieWidgetUI.js
@@ -1,4 +1,4 @@
-import React, { useMemo, useRef, useEffect, useState, useCallback } from 'react';
+import React, { useMemo, useEffect, useState, useCallback } from 'react';
 import PropTypes from 'prop-types';
 import ReactEcharts from '../../custom-components/echarts-for-react';
 import { Box, Grid, Link, styled, useTheme } from '@mui/material';
@@ -6,12 +6,13 @@ import {
   findLargestCategory,
   disableSerie,
   setColor,
-  sortDataDescending,
   calculatePercentage
 } from '../utils/chartUtils';
+import { OTHERS_CATEGORY_NAME, ORDER_TYPES } from '../utils/chartConstants';
 import { processFormatterRes } from '../utils/formatterUtils';
 import PieSkeleton from './components/PieSkeleton';
 import PieCentralText from './components/PieCentralText';
+import usePieCategories from './hooks/usePieCategories';
 import ChartLegend from '../ChartLegend';
 import Typography from '../../components/atoms/Typography';
 
@@ -63,58 +64,8 @@ function PieWidgetUI({
   order
 }) {
   const theme = useTheme();
-  const colorByCategory = useRef({});
   const [selectedItem, setSelectedItem] = useState({});
-  const othersCategory = 'Others';
-
-  // Sort data by size if order is ranking, otherwise keep the original order
-  const orderedData = useMemo(() => {
-    let orderedCategories = [];
-
-    if (order === PieWidgetUI.ORDER_TYPES.RANKING) {
-      orderedCategories = sortDataDescending(data);
-    } else {
-      orderedCategories = [...data];
-    }
-
-    return orderedCategories;
-  }, [data, order]);
-
-  // Limit the number of categories to display, then group the rest into an "Others" category
-  const groupedData = useMemo(() => {
-    let categories = [];
-    let othersValue = 0;
-
-    for (const category of orderedData) {
-      if (categories.length < maxItems) {
-        categories.push({ ...category });
-      } else {
-        othersValue += category.value;
-      }
-    }
-
-    if (othersValue > 0) {
-      categories.push({
-        name: othersCategory,
-        value: othersValue,
-        emphasis: { scale: false }
-      });
-    }
-
-    return categories;
-  }, [maxItems, orderedData]);
-
-  // Add a color to each category
-  const dataWithColor = useMemo(() => {
-    return groupedData.map(processDataItem(colorByCategory, colors, theme));
-  }, [groupedData, colors, theme]);
-
-  // Reset colorByCategory when colors and categories change
-  useEffect(() => {
-    colorByCategory.current = {};
-    // Spread colors array to avoid reference problems
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [...colors, dataWithColor]);
+  const processedData = usePieCategories(data, order, maxItems, colors);
 
   // Tooltip
   const tooltipOptions = useMemo(
@@ -131,17 +82,17 @@ function PieWidgetUI({
 
   // Select the largest category to display in CentralText and calculate its percentage from the total
   const topSelectedCategory = useMemo(() => {
-    if (!dataWithColor || dataWithColor.length === 0) {
+    if (!processedData || processedData.length === 0) {
       return null;
     }
 
     let array;
     if (selectedCategories.length > 0) {
-      array = dataWithColor.filter((dataItem) =>
+      array = processedData.filter((dataItem) =>
         selectedCategories.includes(dataItem.name)
       );
     } else {
-      array = dataWithColor;
+      array = processedData;
     }
 
     const largestCategory = findLargestCategory(array);
@@ -152,7 +103,7 @@ function PieWidgetUI({
     }
 
     let sumValue = 0;
-    for (const category of dataWithColor) {
+    for (const category of processedData) {
       sumValue += category.value;
     }
 
@@ -160,7 +111,7 @@ function PieWidgetUI({
     category.percentage = percentage;
 
     return category;
-  }, [dataWithColor, selectedCategories]);
+  }, [processedData, selectedCategories]);
 
   useEffect(() => {
     if (topSelectedCategory) {
@@ -174,7 +125,7 @@ function PieWidgetUI({
         type: 'pie',
         name,
         animation,
-        data: dataWithColor.map((item) => {
+        data: processedData.map((item) => {
           // Avoid modifying data item
           const clonedItem = { ...item };
 
@@ -208,7 +159,7 @@ function PieWidgetUI({
         }
       }
     ],
-    [name, animation, dataWithColor, labels, selectedCategories, theme]
+    [name, animation, processedData, labels, selectedCategories, theme]
   );
 
   const options = useMemo(
@@ -232,10 +183,10 @@ function PieWidgetUI({
     (params) => {
       if (onSelectedCategoriesChange) {
         const newSelectedCategories = [...selectedCategories];
-        const { name } = dataWithColor[params.dataIndex];
+        const { name } = processedData[params.dataIndex];
 
         // Avoid clicking if the category name is "Others"
-        if (name === othersCategory) {
+        if (name === OTHERS_CATEGORY_NAME) {
           return;
         }
 
@@ -250,7 +201,7 @@ function PieWidgetUI({
         onSelectedCategoriesChange(newSelectedCategories);
       }
     },
-    [dataWithColor, onSelectedCategoriesChange, selectedCategories]
+    [processedData, onSelectedCategoriesChange, selectedCategories]
   );
 
   const handleLegendClick = useCallback(
@@ -305,9 +256,9 @@ function PieWidgetUI({
         />
       </ChartContent>
 
-      {dataWithColor.length > 0 && (
+      {processedData.length > 0 && (
         <ChartLegend
-          series={dataWithColor}
+          series={processedData}
           selectedCategories={selectedCategories}
           onCategoryClick={onSelectedCategoriesChange && handleLegendClick}
         />
@@ -315,15 +266,6 @@ function PieWidgetUI({
     </>
   );
 }
-
-/**
- * Enum for PieWidgetUI order types. 'RANKING' orders the data by value and 'FIXED' keeps the order present in the original data
- * @enum {string}
- */
-PieWidgetUI.ORDER_TYPES = {
-  RANKING: 'ranking',
-  FIXED: 'fixed'
-};
 
 PieWidgetUI.defaultProps = {
   name: null,
@@ -337,7 +279,7 @@ PieWidgetUI.defaultProps = {
   filterable: true,
   selectedCategories: [],
   maxItems: 11,
-  order: PieWidgetUI.ORDER_TYPES.RANKING
+  order: ORDER_TYPES.RANKING
 };
 
 PieWidgetUI.propTypes = {
@@ -360,7 +302,7 @@ PieWidgetUI.propTypes = {
   onSelectedCategoriesChange: PropTypes.func,
   isLoading: PropTypes.bool,
   maxItems: PropTypes.number,
-  order: PropTypes.oneOf(Object.values(PieWidgetUI.ORDER_TYPES))
+  order: PropTypes.oneOf(Object.values(ORDER_TYPES))
 };
 
 export default PieWidgetUI;
@@ -376,21 +318,4 @@ function tooltipFormatter(params) {
     <p style="font-size:12px;font-weight:600;line-height:1.33;margin:4px 0 4px 0;">${params.name}</p>
     <p style="font-size: 12px;font-weight:normal;line-height:1.33;margin:0 0 4px 0;"><span style="${markerStyle}"></span> ${value} (${params.percent}%)</p>
   `.trim();
-}
-
-function processDataItem(colorByCategory, colors, theme) {
-  return (item) => {
-    const { name } = item;
-    const colorUsed = colorByCategory.current[name];
-    if (colorUsed) {
-      item.color = colorUsed;
-    } else {
-      const paletteToUse = colors.length ? colors : theme.palette.qualitative.bold;
-      const colorToUse =
-        paletteToUse[Object.keys(colorByCategory.current).length] || '#fff';
-      colorByCategory.current[name] = colorToUse;
-      item.color = colorToUse;
-    }
-    return item;
-  };
 }
